+++ prob_xcb_dri.c
@@ -364,7 +364,7 @@ int main(int argc, char *argv[])
 
     /* drm events are written in full */
     do {
-     nread = read(id, &vev, 4 * sizeof(struct aux_drm_event_crtc_sq));
+     nread = read((int)id, &vev, 4 * sizeof(struct aux_drm_event_crtc_sq));
      if(nread < 0) {
       if((errno == EAGAIN) || (errno == EWOULDBLOCK)) {
        break;
